refactor(mealplan): split MealPlanEndpoints into slot-ops partial + colocate handlers
MealPlanEndpoints was 280 lines bundling 14 endpoints with handlers
scattered: six of them sat after the map block at the bottom of the
file rather than next to their MapPost / MapPut calls, which the
CLAUDE.md endpoint convention asks for explicitly.

Two changes:

- Split slot-mutation endpoints (assign / adjust-servings /
  cook-with-leftovers / swap / confirm / clear) into a new
  MealPlanEndpoints.Slots.cs partial exposed via a private
  MapSlotEndpoints(this RouteGroupBuilder) extension. Main file keeps
  plan-level reads (get / planned-recipes / extended-mode /
  generate-shopping-list / copy-to-week) and history search.
- Every handler now sits as a `static async Task<IResult>` local
  function immediately after its route definition.

- MealPlanEndpoints.cs: 280 -> 131 lines
- MealPlanEndpoints.Slots.cs: 166 lines (new)
